requirements.txt
numpy
scipy
pyrtlsdr
pyaudio
